I cooked So HARD -Added En Passant -Added Casteling -Added Promotions -I also moved around a bunch of stuff and added charToPiece in board.cpp
A bit of syntax.
current promtion syntax is ex e7e8Q which means pawn on e7 move to e8 and promotes to queen, atm the Promotion piece symbol has to be a capital letter.

En passant works as a normal capture lets say last move was pawn d7d5 for black and white has pawn on e5, e5d6 would be perfectly legal since of en passant and yes it would take away black pawn on d5.

Castleling works with O-O for short castle and O-O-O for long and this syntax works for both players.

Everything problably has bugs so we arent done quite yet plus the code is ultra sloppy cause I just wanted it done ;-;

Also I might have missed something that I did but rn I cant think of any more things that I did.
oh I did a bit of a cheeky thing which is that the casteling is both checked and handle in makemove not islegal and promotions is semi handled in make move and islegal while enpassant is almost fully handled in islegal ._. this is very stökigt yes swenglish.